+++ R/type-data-frame.R
@@ -184,11 +184,10 @@ vec_ptype_full.data.frame <- function(x, ...) {
 #' @export
 vec_ptype_abbr.data.frame <- function(x, ...) {
   if (inherits_only(x, "data.frame")) {
-    abbr <- "df"
+    "df"
   } else {
-    abbr <- class(x)[[1]]
+    class(x)[[1]]
   }
-  paste0(abbr, vec_ptype_shape(x))
 }
 
 #' @export