+++ tests/testthat/test-ptype-abbr-full.R
@@ -20,8 +20,14 @@ test_that("non objects default to type + shape", {
 
 })
 
+test_that("non objects can omit shape", {
+  expect_equal(vec_ptype_abbr(ones(10), suffix_shape = FALSE), "dbl")
+  expect_equal(vec_ptype_abbr(ones(0, 10), suffix_shape = FALSE), "dbl")
+  expect_equal(vec_ptype_abbr(ones(10, 0), suffix_shape = FALSE), "dbl")
+})
+
 test_that("objects default to first class", {
-  x <- structure(1, class = "foofy")
+  x <- structure(1, class = c("foofy", "goofy"))
   expect_equal(vec_ptype_full(x), "foofy")
   expect_equal(vec_ptype_abbr(x), "foofy")
 })
@@ -38,6 +44,26 @@ test_that("complex and factor as expected (#323)", {
   expect_equal(vec_ptype_abbr(factor()), "fct")
 })
 
-test_that("named lists are tagged (#322)", {
+test_that("named lists are always tagged (#322)", {
   expect_identical(vec_ptype_abbr(list(x = 1, y = 2)), "named list")
+  expect_identical(vec_ptype_abbr(list(x = 1, y = 2), prefix_named = TRUE), "named list")
+})
+
+test_that("named atomics are tagged optionally (#781)", {
+  expect_identical(vec_ptype_abbr(c(x = 1, y = 2), prefix_named = TRUE), "named dbl")
+  expect_identical(vec_ptype_abbr(c(x = 1L, y = 2L), prefix_named = TRUE), "named int")
+})
+
+test_that("vec_ptype_abbr() adds named tag in case of row names", {
+  expect_equal(
+    vec_ptype_abbr(mtcars, prefix_named = TRUE),
+    "named df[,11]"
+  )
+
+  mat <- matrix(1:4, 2)
+  rownames(mat) <- c("foo", "bar")
+  expect_equal(
+    vec_ptype_abbr(mat, prefix_named = TRUE),
+    "named int[,2]"
+  )
 })