+++ tests/testthat/test-type-table.R
@@ -6,7 +6,7 @@ test_that("ptype print methods are descriptive", {
   tab2 <- new_table(dim = c(0L, 1L, 2L, 1L))
 
   expect_equal(vec_ptype_abbr(tab1), "table")
-  expect_equal(vec_ptype_abbr(tab2), "table")
+  expect_equal(vec_ptype_abbr(tab2), "table[,1,2,1]")
 
   expect_equal(vec_ptype_full(tab1), "table")
   expect_equal(vec_ptype_full(tab2), "table[,1,2,1]")